lib: make lazyDOMException more common
