+++ lib/buffer.js
@@ -75,6 +75,7 @@ const {
 const {
   customInspectSymbol,
   isInsideNodeModules,
+  lazyDOMException,
   normalizeEncoding,
   kIsEncodingSymbol
 } = require('internal/util');
@@ -1208,13 +1209,9 @@ if (internalBinding('config').hasIntl) {
   };
 }
 
-let DOMException;
-
-const lazyInvalidCharError = hideStackFrames((message, name) => {
-  if (DOMException === undefined)
-    DOMException = internalBinding('messaging').DOMException;
-  throw new DOMException('Invalid character', 'InvalidCharacterError');
-});
+const invalidCharError = () => {
+  return lazyDOMException('Invalid character', 'InvalidCharacterError');
+};
 
 function btoa(input) {
   // The implementation here has not been performance optimized in any way and
@@ -1223,7 +1220,7 @@ function btoa(input) {
   input = `${input}`;
   for (let n = 0; n < input.length; n++) {
     if (input[n].charCodeAt(0) > 0xff)
-      lazyInvalidCharError();
+      invalidCharError();
   }
   const buf = Buffer.from(input, 'latin1');
   return buf.toString('base64');
@@ -1239,7 +1236,7 @@ function atob(input) {
   input = `${input}`;
   for (let n = 0; n < input.length; n++) {
     if (!kBase64Digits.includes(input[n]))
-      lazyInvalidCharError();
+      invalidCharError();
   }
   return Buffer.from(input, 'base64').toString('latin1');
 }
