+++ lib/internal/crypto/aes.js
@@ -36,7 +36,6 @@ const {
   getArrayBufferOrView,
   hasAnyNotIn,
   jobPromise,
-  lazyDOMException,
   validateByteLength,
   validateKeyOps,
   validateMaxBufferLength,
@@ -44,6 +43,10 @@ const {
   kHandle,
   kKeyObject,
 } = require('internal/crypto/util');
+
+const {
+  lazyDOMException,
+} = require('internal/util');
 
 const { PromiseReject } = primordials;
