+++ lib/internal/crypto/diffiehellman.js
@@ -48,6 +48,10 @@ const {
 } = require('internal/util/types');
 
 const {
+  lazyDOMException,
+} = require('internal/util');
+
+const {
   KeyObject,
   InternalCryptoKey,
   createPrivateKey,
@@ -66,7 +70,6 @@ const {
   getUsagesUnion,
   hasAnyNotIn,
   jobPromise,
-  lazyDOMException,
   toBuf,
   kHandle,
   kKeyObject,
