+++ lib/internal/crypto/dsa.js
@@ -44,12 +44,15 @@ const {
   getUsagesUnion,
   hasAnyNotIn,
   jobPromise,
-  lazyDOMException,
   normalizeHashName,
   validateKeyOps,
   kKeyObject,
   kHandle,
 } = require('internal/crypto/util');
+
+const {
+  lazyDOMException,
+} = require('internal/util');
 
 function verifyAcceptableDsaKeyUse(name, type, usages) {
   let checkSet;
