+++ lib/internal/crypto/ec.js
@@ -38,13 +38,16 @@ const {
   getUsagesUnion,
   hasAnyNotIn,
   jobPromise,
-  lazyDOMException,
   normalizeHashName,
   validateKeyOps,
   kHandle,
   kKeyObject,
   kNamedCurveAliases,
 } = require('internal/crypto/util');
+
+const {
+  lazyDOMException,
+} = require('internal/util');
 
 const {
   generateKeyPair,
