+++ lib/internal/crypto/hkdf.js
@@ -23,7 +23,6 @@ const { kMaxLength } = require('buffer');
 
 const {
   getArrayBufferOrView,
-  lazyDOMException,
   normalizeHashName,
   toBuf,
   validateByteSource,
@@ -34,6 +33,10 @@ const {
   createSecretKey,
   isKeyObject,
 } = require('internal/crypto/keys');
+
+const {
+  lazyDOMException,
+} = require('internal/util');
 
 const {
   isAnyArrayBuffer,
