+++ lib/internal/crypto/mac.js
@@ -18,13 +18,16 @@ const {
   getHashLength,
   hasAnyNotIn,
   jobPromise,
-  lazyDOMException,
   normalizeHashName,
   validateBitLength,
   validateKeyOps,
   kHandle,
   kKeyObject,
 } = require('internal/crypto/util');
+
+const {
+  lazyDOMException,
+} = require('internal/util');
 
 const {
   codes: {
