+++ lib/internal/crypto/pbkdf2.js
@@ -25,10 +25,13 @@ const { ERR_MISSING_OPTION } = require('internal/errors').codes;
 const {
   getArrayBufferOrView,
   getDefaultEncoding,
-  lazyDOMException,
   normalizeHashName,
   kKeyObject,
 } = require('internal/crypto/util');
+
+const {
+  lazyDOMException,
+} = require('internal/util');
 
 function pbkdf2(password, salt, iterations, keylen, digest, callback) {
   if (typeof digest === 'function') {
