+++ lib/internal/crypto/random.js
@@ -27,7 +27,7 @@ const {
 
 const {
   lazyDOMException,
-} = require('internal/crypto/util');
+} = require('internal/util');
 
 const { Buffer, kMaxLength } = require('buffer');
