+++ lib/internal/crypto/rsa.js
@@ -40,13 +40,16 @@ const {
   getUsagesUnion,
   hasAnyNotIn,
   jobPromise,
-  lazyDOMException,
   normalizeHashName,
   validateKeyOps,
   validateMaxBufferLength,
   kHandle,
   kKeyObject,
 } = require('internal/crypto/util');
+
+const {
+  lazyDOMException,
+} = require('internal/util');
 
 const {
   isUint8Array,
