+++ lib/internal/crypto/scrypt.js
@@ -30,9 +30,12 @@ const {
 const {
   getArrayBufferOrView,
   getDefaultEncoding,
-  lazyDOMException,
   kKeyObject,
 } = require('internal/crypto/util');
+
+const {
+  lazyDOMException,
+} = require('internal/util');
 
 const defaults = {
   N: 16384,
