+++ lib/internal/crypto/util.js
@@ -51,6 +51,7 @@ const { Buffer } = require('buffer');
 const {
   cachedResult,
   filterDuplicateStrings,
+  lazyDOMException,
 } = require('internal/util');
 
 const {
@@ -69,13 +70,6 @@ function lazyRequire(name) {
     ret = lazyRequireCache[name] = require(name);
   return ret;
 }
-
-let DOMException;
-const lazyDOMException = hideStackFrames((message, name) => {
-  if (DOMException === undefined)
-    DOMException = internalBinding('messaging').DOMException;
-  return new DOMException(message, name);
-});
 
 var defaultEncoding = 'buffer';
 
@@ -428,7 +422,6 @@ module.exports = {
   normalizeAlgorithm,
   normalizeHashName,
   hasAnyNotIn,
-  lazyDOMException,
   validateBitLength,
   validateByteLength,
   validateByteSource,
