+++ lib/internal/crypto/webcrypto.js
@@ -49,7 +49,6 @@ const {
 const {
   getArrayBufferOrView,
   hasAnyNotIn,
-  lazyDOMException,
   lazyRequire,
   normalizeAlgorithm,
   normalizeHashName,
@@ -58,6 +57,10 @@ const {
   kHandle,
   kKeyObject,
 } = require('internal/crypto/util');
+
+const {
+  lazyDOMException,
+} = require('internal/util');
 
 const {
   getRandomValues,
