+++ lib/internal/fs/promises.js
@@ -76,7 +76,7 @@ const {
   validateInteger,
 } = require('internal/validators');
 const pathModule = require('path');
-const { promisify } = require('internal/util');
+const { lazyDOMException, promisify } = require('internal/util');
 const { EventEmitterMixin } = require('internal/event_target');
 const { watch } = require('internal/fs/watchers');
 const { isIterable } = require('internal/streams/utils');
@@ -211,8 +211,7 @@ class FileHandle extends EventEmitterMixin(JSTransferable) {
 
   [kTransfer]() {
     if (this[kClosePromise] || this[kRefs] > 1) {
-      const DOMException = internalBinding('messaging').DOMException;
-      throw new DOMException('Cannot transfer FileHandle while in use',
+      throw lazyDOMException('Cannot transfer FileHandle while in use',
                              'DataCloneError');
     }
