+++ lib/internal/util.js
@@ -28,6 +28,7 @@ const {
 } = primordials;
 
 const {
+  hideStackFrames,
   codes: {
     ERR_INVALID_ARG_TYPE,
     ERR_NO_CRYPTO,
@@ -441,6 +442,13 @@ function createDeferredPromise() {
   return { promise, resolve, reject };
 }
 
+let DOMException;
+const lazyDOMException = hideStackFrames((message, name) => {
+  if (DOMException === undefined)
+    DOMException = internalBinding('messaging').DOMException;
+  return new DOMException(message, name);
+});
+
 module.exports = {
   assertCrypto,
   cachedResult,
@@ -457,6 +465,7 @@ module.exports = {
   isError,
   isInsideNodeModules,
   join,
+  lazyDOMException,
   normalizeEncoding,
   once,
   promisify,
